+++ client/src/lib.rs
@@ -1,4 +1,0 @@
-mod error;
-pub mod http;
-mod models;
-mod views;