+++ CHANGELOG.md
@@ -4,6 +4,12 @@ This file documents all notable changes to this project. The format of this file
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Removed
+
+- The `client` subcommand was obsoleted by REsolutions-frontend.
+
 ## [0.6.2] - 2019-11-08
 
 ### Added
@@ -186,6 +192,7 @@ v0.4.1
   - `ETCD_ADDR`
   - `ETCD_SIG_KEY`
 
+[Unreleased]: https://gitlab.com/resolutions/review/compare/0.6.2...master
 [0.6.2]: https://gitlab.com/resolutions/review/compare/0.6.1...0.6.2
 [0.6.1]: https://gitlab.com/resolutions/review/compare/0.6.0...0.6.1
 [0.6.0]: https://gitlab.com/resolutions/review/compare/0.5.0...0.6.0